fix(mobile): refactor feature flags, gate collectibles and waitlist, ref ENG-16 PRO-12
