+++ apps/mobile/src/app/_layout.tsx
@@ -12,8 +12,7 @@ import { HapticsProvider } from '@/core/haptics-provider';
 import { LeatherQueryProvider } from '@/core/leather-query-provider';
 import { ThemeProvider } from '@/core/theme-provider';
 import { BrowserSheet } from '@/features/browser/browser/browser-sheet';
-import { featureFlagClient } from '@/features/feature-flags/feature-flag';
-import { setupFeatureFlags } from '@/features/feature-flags/setup-feature-flags';
+import { featureFlagClient, setupFeatureFlags } from '@/features/feature-flags';
 import { useWatchNotificationAddresses } from '@/features/notifications/use-notifications';
 import { ReceiveSheet } from '@/features/receive/receive-sheet';
 import { SendSheet } from '@/features/send/send-sheet';
