+++ apps/mobile/src/app/account/[account]/account.layout.tsx
@@ -3,6 +3,7 @@ import { PageLayout } from '@/components/page/page.layout';
 import { BalancesWidget } from '@/components/widgets/balances/balances-widget';
 import { CollectiblesWidget } from '@/components/widgets/collectibles/collectibles-widget';
 import { AccountBalances } from '@/features/balances/balances';
+import { useCollectiblesFlag } from '@/features/feature-flags';
 import { useAccountAddresses } from '@/hooks/use-account-addresses';
 import { useAccountActivityQuery } from '@/queries/activity/account-activity.query';
 import { useAccountCollectibles } from '@/queries/collectibles/account-collectibles.query';
@@ -24,6 +25,7 @@ export function AccountLayout({ account, balance }: AccountLayoutProps) {
   const accountAddresses = useAccountAddresses(account.fingerprint, account.accountIndex);
   const { data: activity, isLoading } = useAccountActivityQuery(accountAddresses);
   const collectibles = useAccountCollectibles(account.fingerprint, account.accountIndex);
+  const releaseCollectibles = useCollectiblesFlag();
   return (
     <PageLayout>
       <AccountOverview
@@ -59,15 +61,17 @@ export function AccountLayout({ account, balance }: AccountLayoutProps) {
         />
       )}
 
-      <CollectiblesWidget
-        collectibles={collectibles}
-        onPressHeader={() =>
-          router.navigate({
-            pathname: AppRoutes.AccountCollectibles,
-            params: { accountId: account.id, accountName: account.name },
-          })
-        }
-      />
+      {collectibles && releaseCollectibles && (
+        <CollectiblesWidget
+          collectibles={collectibles}
+          onPressHeader={() =>
+            router.navigate({
+              pathname: AppRoutes.AccountCollectibles,
+              params: { accountId: account.id, accountName: account.name },
+            })
+          }
+        />
+      )}
     </PageLayout>
   );
 }
