--- apps/mobile/src/features/feature-flags/feature-flag.ts
@@ -1,37 +1,0 @@
-import { isProduction } from '@/shared/environment';
-import { AutoEnvAttributes, ReactNativeLDClient } from '@launchdarkly/react-native-client-sdk';
-import * as Application from 'expo-application';
-
-export const featureFlagClient = new ReactNativeLDClient(
-  // TODO: do not fallback to empty string
-  process.env.EXPO_PUBLIC_LAUNCH_DARKLY ?? '',
-  AutoEnvAttributes.Enabled,
-  {
-    debug: !isProduction(),
-    applicationInfo: {
-      id: 'leather-mobile-wallet',
-      version: Application.nativeApplicationVersion ?? '0.0.1',
-    },
-  }
-);
-
-export const featureFlagKeys = {
-  releasePushNotifications: {
-    key: 'release_push_notifications',
-    defaultValue: false,
-  },
-  releaseBrowserFeature: {
-    key: 'release_browser_feature',
-    defaultValue: false,
-  },
-  releaseLocaleFeature: {
-    key: 'release_locale_feature',
-    defaultValue: false,
-  },
-  releaseNotificationsFeature: {
-    key: 'release_notifications_feature',
-    defaultValue: false,
-  },
-};
-
-export type FEATURE_FLAG_KEYS = keyof typeof featureFlagKeys;
