+++ apps/mobile/src/features/send/components/recipient/recipient.tsx
@@ -1,6 +1,6 @@
 import { FieldError } from 'react-hook-form';
 
-import { useNewRecipientFlowFeatureFlag } from '@/features/feature-flags/use-feature-flags';
+import { useNewRecipientFlowFlag } from '@/features/feature-flags';
 import { Recipient as RecipientV1 } from '@/features/send/components/recipient/v1/recipient.v1';
 import { Recipient as RecipientV2 } from '@/features/send/components/recipient/v2/recipient.v2';
 import { ZodSchema } from 'zod';
@@ -17,6 +17,6 @@ interface RecipientProps {
 }
 
 export function Recipient(props: RecipientProps) {
-  const RecipientComponent = useNewRecipientFlowFeatureFlag() ? RecipientV2 : RecipientV1;
+  const RecipientComponent = useNewRecipientFlowFlag() ? RecipientV2 : RecipientV1;
   return <RecipientComponent {...props} />;
 }
